+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/AbstractAttributeSetter.php
@@ -47,14 +47,6 @@ abstract class AbstractAttributeSetter implements AttributeSetterInterface
     /**
      * {@inheritdoc}
      */
-    public function supports(AttributeInterface $attribute)
-    {
-        return $this->supportsAttribute($attribute);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function supportsAttribute(AttributeInterface $attribute)
     {
         return in_array($attribute->getAttributeType(), $this->supportedTypes);