+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/AbstractFieldSetter.php
@@ -2,8 +2,6 @@
 
 namespace Pim\Bundle\CatalogBundle\Updater\Setter;
 
-use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
-
 /**
  * Abstract field setter
  *
@@ -16,27 +14,11 @@ abstract class AbstractFieldSetter implements FieldSetterInterface
     /** @var array */
     protected $supportedFields = [];
 
-    public function setValue(array $products, AttributeInterface $attribute, $data, $locale = null, $scope = null)
-    {
-        throw new \Exception('This method is not supported for field setters');
-
-    }
-
     /**
      * {@inheritdoc}
      */
     public function supportsField($field)
     {
         return in_array($field, $this->supportedFields);
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @deprecated will be removed in 1.5, use methods from FieldSetterInterface and AttributeSetterInterface
-     */
-    public function supports(AttributeInterface $attribute)
-    {
-        return false;
     }
 }