+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/BooleanAttributeSetter.php
@@ -33,20 +33,6 @@ class BooleanAttributeSetter extends AbstractAttributeSetter
 
     /**
      * {@inheritdoc}
-     *
-     * Expected data input format : true|false
-     *
-     * @deprecated will be removed in 1.5, use method setAttributeData
-     */
-    public function setValue(array $products, AttributeInterface $attribute, $data, $locale = null, $scope = null)
-    {
-        foreach ($products as $product) {
-            $this->setAttributeData($product, $attribute, $data, ['locale' => $locale, 'scope' => $scope]);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
      */
     public function setAttributeData(
         ProductInterface $product,