+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/DateAttributeSetter.php
@@ -34,18 +34,6 @@ class DateAttributeSetter extends AbstractAttributeSetter
     /**
      * {@inheritdoc}
      *
-     * @deprecated will be removed in 1.5, use method setAttributeData
-     */
-    public function setValue(array $products, AttributeInterface $attribute, $data, $locale = null, $scope = null)
-    {
-        foreach ($products as $product) {
-            $this->setAttributeData($product, $attribute, $data, ['locale' => $locale, 'scope' => $scope]);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     *
      * Expected data input format : "yyyy-mm-dd"
      */
     public function setAttributeData(