+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/MediaAttributeSetter.php
@@ -56,18 +56,6 @@ class MediaAttributeSetter extends AbstractAttributeSetter
     /**
      * {@inheritdoc}
      *
-     * @deprecated will be removed in 1.5, use method setAttributeData
-     */
-    public function setValue(array $products, AttributeInterface $attribute, $data, $locale = null, $scope = null)
-    {
-        foreach ($products as $product) {
-            $this->setAttributeData($product, $attribute, $data, ['locale' => $locale, 'scope' => $scope]);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     *
      * Expected data input format :
      * {
      *     "originalFilename": "original_filename.extension",