+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/MultiSelectAttributeSetter.php
@@ -41,18 +41,6 @@ class MultiSelectAttributeSetter extends AbstractAttributeSetter
     /**
      * {@inheritdoc}
      *
-     * @deprecated will be removed in 1.5, use method setAttributeData
-     */
-    public function setValue(array $products, AttributeInterface $attribute, $data, $locale = null, $scope = null)
-    {
-        foreach ($products as $product) {
-            $this->setAttributeData($product, $attribute, $data, ['locale' => $locale, 'scope' => $scope]);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     *
      * Expected data input format: ["option_code", "other_option_code"]
      */
     public function setAttributeData(