+++ src/Pim/Bundle/CatalogBundle/Updater/Setter/NumberAttributeSetter.php
@@ -35,18 +35,6 @@ class NumberAttributeSetter extends AbstractAttributeSetter
     /**
      * {@inheritdoc}
      *
-     * @deprecated will be removed in 1.5, use method setAttributeData
-     */
-    public function setValue(array $products, AttributeInterface $attribute, $data, $locale = null, $scope = null)
-    {
-        foreach ($products as $product) {
-            $this->setAttributeData($product, $attribute, $data, ['locale' => $locale, 'scope' => $scope]);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     *
      * Expected data input format: "12.0"|"12"|12|12.3
      */
     public function setAttributeData(